+++ src/FsHotWatch.Cli/InitConfig.fs
@@ -109,6 +109,7 @@ let generateConfig (projectPaths: string list) : DaemonConfiguration =
                    CoverageDir = "coverage"
                    DependsOn = [] |}
       FileCommands = []
+      Preprocessors = []
       Coverage = None
       Exclude = []
       IncludeOutsideRepo = false